[ELF][MIPS] Don't change every dynamic relocation into R_MIPS_REL32
I was debugging some strange crashes in a dynamically linked CHERI binary
and it turns out that LLD always emits R_MIPS_REL32/R_MIPS_64 dynamic
relocations (even for an undefined symbol). The crash I was getting
happened because the program loaded a global function pointer referencing
an undefined symbol. However, because the wrong relocation was being
emitted RTLD only added the relocbase to the function pointer. Therefore
the program crashed when jumping to the start of the mapped ELF file
since that does not contain valid instructions.