Allow implicit conversion of address-of expressions to capabilities in the hybrid ABI. This allows patterns of the form:
int foo;
int * __capability bar;
bar = &foo;

It will also allow assigning to a void * __capability but will otherwise output
an error if the types don't match.

I'm not sure if implicit conversion should happen when taking the address of a
function so have let it output an error (in the hybrid ABI):

void f() { ... }

void (* __capability fp)() = &f;

If it makes sense for this to be allowed then we can enable it.

This implicit conversion works for both C and C++. Implicit conversion of
address-of member expressions also work.

Fixes #182.