Add a test case for the broken QReadWriteLock
We are currently generating a $ddc derived capability for the dummy value
stored in QReadWriteLock to mark it as locked for read/write. This value
is then later compared to a null derived `(__uintcap_t)2` and returns false
so the code assumes it is a valid pointer. This clearly cannot work and
crashes horribly.